Fix the ImplicitArrayLowerBound canonicalizer rule
