+++ src/canonicalizer/CMakeLists.txt
@@ -12,6 +12,7 @@ add_library(sourcemeta_jsonbinpack_canonicalizer
   rules/drop_non_boolean_keywords_validation.h
   rules/equal_numeric_bounds_as_const.h
   rules/if_without_then_else.h
+  rules/implicit_array_lower_bound.h
   rules/max_contains_without_contains.h
   rules/min_contains_without_contains.h
   rules/null_as_const.h
