+++ src/canonicalizer/rules/implicit_array_lower_bound.h
@@ -1,0 +1,30 @@
+namespace sourcemeta::jsonbinpack::canonicalizer {
+
+class ImplicitArrayLowerBound final : public sourcemeta::alterschema::Rule {
+public:
+  ImplicitArrayLowerBound() : Rule("implicit_array_lower_bound"){};
+  [[nodiscard]] auto
+  condition(const sourcemeta::jsontoolkit::Value &schema,
+            const std::string &dialect,
+            const std::unordered_map<std::string, bool> &vocabularies) const
+      -> bool override {
+    return dialect == "https://json-schema.org/draft/2020-12/schema" &&
+           vocabularies.contains(
+               "https://json-schema.org/draft/2020-12/vocab/validation") &&
+           sourcemeta::jsontoolkit::is_object(schema) &&
+           sourcemeta::jsontoolkit::defines(schema, "type") &&
+           sourcemeta::jsontoolkit::is_string(
+               sourcemeta::jsontoolkit::at(schema, "type")) &&
+           sourcemeta::jsontoolkit::to_string(
+               sourcemeta::jsontoolkit::at(schema, "type")) == "array" &&
+           !sourcemeta::jsontoolkit::defines(schema, "minItems");
+  }
+
+  auto transform(sourcemeta::jsontoolkit::JSON &document,
+                 sourcemeta::jsontoolkit::Value &value) const -> void override {
+    sourcemeta::jsontoolkit::assign(document, value, "minItems",
+                                    sourcemeta::jsontoolkit::from(0));
+  }
+};
+
+} // namespace sourcemeta::jsonbinpack::canonicalizer
